+++ src/Keter/Types/V04.hs
@@ -7,12 +7,15 @@ import           Control.Applicative
 import           Data.Aeson
 import           Data.Bool
 import           Data.Conduit.Network              (HostPreference)
-import qualified Data.Set                          as Set
 import           Data.String                       (fromString)
 import           Keter.Yaml.FilePath
 import qualified System.FilePath                   as F
 import           Keter.Types.Common
 import           Keter.Rewrite(ReverseProxyConfig)
+import           Data.Text                  (Text)
+import           System.FilePath            (FilePath)
+import           Data.Set                   (Set)
+import qualified Data.Set                   as Set
 import qualified Network.Wai.Handler.Warp          as Warp
 import qualified Network.Wai.Handler.WarpTLS       as WarpTLS
 import qualified Network.TLS.SessionManager        as TLSSession